src/utils/unifiedIntelligenceField.ts
import { LRUCache } from 'lru-cache';

export class UnifiedIntelligenceField {
  private static instance: UnifiedIntelligenceField;
  private cache: LRUCache<string, any>;

  private constructor() {
    this.cache = new LRUCache({
      max: 100,
      maxAge: 1000 * 60 * 60 // 1 hour
    });
  }

  public static getInstance(): UnifiedIntelligenceField {
    if (!UnifiedIntelligenceField.instance) {
      UnifiedIntelligenceField.instance = new UnifiedIntelligenceField();
    }
    return UnifiedIntelligenceField.instance;
  }

  async getTrendingTopics(): Promise<any[]> {
    const cacheKey = 'trendingTopics';
    const cachedTopics = this.cache.get(cacheKey);
    if (cachedTopics) {
      return cachedTopics;
    }

    // Simulate API call or complex computation
    await new Promise(resolve => setTimeout(resolve, 1000));
    const topics = [
      { name: "Sustainable Fashion", tweet_volume: 10000 },
      { name: "AI Ethics", tweet_volume: 8000 },
      { name: "Remote Work", tweet_volume: 15000 }
    ];

    this.cache.set(cacheKey, topics);
    return topics;
  }

  // ... other methods with similar caching strategy ...
}